+++ src/gui/models/types.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2024 Team Dissolve and contributors
 
-#include "gui/types.h"
+#include "gui/models/types.h"
 #include "gui/models/addForcefieldDialogModel.h"
 #include "gui/models/atomTypeModel.h"
 #include "gui/models/configurationModel.h"
