+++ src/qml-gui/qml/DissolveMain.qml
@@ -1,0 +1,13 @@
+import QtQuick
+import QtQuick.Controls
+import QtQuick.Layouts
+
+ApplicationWindow {
+    id: mainWindow
+    visible: true
+    width: 600
+    height: 400
+    title: "Dissolve"
+
+
+}
